pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uasp"
version = "0.1.0"
description = "Unified Agent Skills Protocol - A structured, machine-readable format for agent skills"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "UASP Contributors" }
]
keywords = ["uasp", "agent", "skills", "llm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "click>=8.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
all = ["uasp[llm,dev]"]

[project.scripts]
uasp = "uasp.cli.main:cli"

[project.urls]
Homepage = "https://github.com/uasp/uasp-py"
Documentation = "https://uasp.dev"
Repository = "https://github.com/uasp/uasp-py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/uasp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
